InWork(desafio): Front - Alterado todo o front para Reacj.js - Criado a tela que listará todas as pautas, podendo votar, deletar ou criar novas pautas - Criado as APIs para enviar os dados para o backend - No backend ajustado todos os fluxos para poder salvar as novas pautas, aberturar de sessões e salvar os votos - Criado o fluxo de validação se o associado pode ou não votar em determinada pauta - Criado um scheduler para verificar quando uma sessão deve ser fechada.

+++ src/main/java/com/desafiofullstask/votacao/service/SessionService.java
@@ -24,7 +24,7 @@ public class SessionService {
         sessionRepository.deleteById(id);
     }
 
-    public void changeStatus(Integer id, Long status) {
+    public void changeStatus(Integer id, int status) {
         Session session = sessionRepository.findById(id).orElse(null);
         if (session != null) {
             session.setStatus(status);
